backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
supabase==1.0.0
# Use a huggingface-hub version available on PyPI for this environment
huggingface-hub==0.18.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.11
